Move crate::field::test_util to crate::
Since I need similar testing utilities in filter::
